Brightness adjuster program
This program takes a battery percentage and a brightness percentage, when your computer reaches that battery percentage, your screen is set to that brightness percentage. Unfortunately, it only works on Windows and Linux, but not MacOS. It is my most advanced project yet and it took me a while, however, it is still pretty janky, for example, you can't close the command prompt if you want it to work. I hope to make more advanced projects like this in the future.